test(base,jsii,githubaction): removed test duplicates
