+++ test/base/commitlint.ts
@@ -1,0 +1,69 @@
+import { SynthOutput } from 'projen/lib/util/synth';
+import * as common from './common';
+import { TaskSteps } from '../../src/types';
+
+/**
+ * Validates that commitlintrc template matches expected template.
+ * @param snapshot Synthesized project output.
+ */
+export function testRcTemplate(snapshot: SynthOutput): void {
+  const expectedTemplateLines: string = [
+    '// Generated by projen.To modify, edit.projenrc.ts and run "npx projen".',
+    '',
+    "import type { UserConfig } from '@commitlint/types';",
+    "import { RuleConfigSeverity } from '@commitlint/types';",
+    '',
+    'const Configuration: UserConfig = {',
+    "  extends: ['@commitlint/config-conventional'],",
+    '  rules: {',
+    "    'type-enum': [",
+    '      RuleConfigSeverity.Error,',
+    "      'always',",
+    "      ['build', 'chore', 'ci', 'docs', 'feat', 'fix', 'perf', 'refactor', 'revert', 'style', 'test', 'delete'],",
+    '    ],',
+    "    'scope-empty': [RuleConfigSeverity.Error, 'never'],",
+    "    'subject-empty': [RuleConfigSeverity.Error, 'never'],",
+    '  },',
+    "  helpUrl: 'https://github.com/conventional-changelog/commitlint/#what-is-commitlint',",
+    '};',
+    '',
+    'export default Configuration;',
+  ].join('\n');
+  expect(snapshot['.commitlintrc.ts']).toBe(expectedTemplateLines);
+}
+
+/**
+ * Validates that npm scripts are added properly.
+ * @param snapshot Synthesized project output.
+ */
+export function testScripts(snapshot: SynthOutput): void {
+  const expectedTasks: TaskSteps = {
+    commit: ['commit'],
+  };
+  common.testScripts(snapshot, expectedTasks);
+}
+
+/**
+ * Validates that npm dev dependencies are added properly.
+ * @param snapshot Synthesized project output.
+ */
+export function testDevDependencies(snapshot: SynthOutput): void {
+  const expectedDevDependencies: string[] = [
+    '@commitlint/cli',
+    '@commitlint/config-conventional',
+    '@commitlint/prompt-cli',
+    '@commitlint/types',
+    'lint-staged',
+  ];
+  common.testDevDependencies(snapshot, expectedDevDependencies);
+}
+
+/**
+ * Validates that project related files are added to .gitattributes and defined as linguist-generated.
+ * @param snapshot Synthesized project output.
+ */
+export function testGitAttributes(snapshot: SynthOutput): void {
+  const patterns: RegExp[] = [/\/\.commitlintrc\.ts linguist-generated( $|\s|$)/m];
+
+  common.testGitAttributes(snapshot, patterns);
+}
